Extract tests with local HTTP/HTTPS server from TestGemRemoteFetcher
